Implement correct frontend flow with guest credit tracking
Backend:
- model/guest.go: Credit int -> float64 to match Lua float arithmetic in Redis
- main.go: add AllowCredentials:true to CORS (required for cross-origin cookie)

Frontend:
- GuestContext.tsx: new client context that fetches /guest/status on mount,
  exposes credit count and refreshCredit() to the whole app
- layout.tsx: wrap app in GuestProvider so Navbar and pages share credit state
- Navbar: show live credit count; "เครดิตหมด" in red when 0
- page.tsx: add credentials:"include" on fetch; distinct error messages per
  HTTP status (402 -> insufficient credit, 401 -> session expired, 500 -> retry);
  call refreshCredit() after each successful translation
- second/page.tsx: remove dead result.translated fallback, use result.result only

+++ apps/api/internal/model/guest.go
@@ -4,6 +4,6 @@ import "time"
 
 type Guest struct {
 	ID        string    `json:"id"`
-	Credit    int       `json:"credit"`
+	Credit    float64   `json:"credit"`
 	CreatedAt time.Time `json:"created_at"`
 }
